Move versioning module to app package to prevent import loop
app/api/__init__ was causing an import loop when versioning was in the
api package. Since the main (potential) user of this is the utils
package which is directly under app anyways, I'm assuming this won't
present an issue.
